NullPointerException when the Tag.LUTData is null
Hello Nicolas, I propose you to make this modification, because we happened to receive a non-standard dicom which contained the tag (0028,3010) and the (0028,3006) was not present. (0028,3006) is of type 1 if (0028,3010) is present. 

It happend in the case when the line 139 bData is null:

bData = dicomLutObject.getBytes(Tag.LUTData); 

So, at line 146 when I do bData.length I get a NullPointerException.

I suggest you add this code to solve the problem.